support installing from a specific revision on remote

+++ workbench/src/main/setupAddRemovePlugin.js
@@ -52,32 +52,55 @@ function spawnWithLogging(cmd, args, options) {
 export function setupAddPlugin() {
   ipcMain.handle(
     ipcMainChannels.ADD_PLUGIN,
-    async (e, pluginURL) => {
+    async (e, url, revision, path) => {
       try {
-        logger.info('adding plugin at', pluginURL);
-        const micromamba = settingsStore.get('micromamba');
-        // Create a temporary directory and check out the plugin's pyproject.toml
-        const tmpPluginDir = fs.mkdtempSync(upath.join(tmpdir(), 'natcap-invest-'));
-        await spawnWithLogging(
-          'git',
-          ['clone', '--depth', '1', '--no-checkout', pluginURL, tmpPluginDir]
-        );
-        await spawnWithLogging(
-          'git',
-          ['checkout', 'HEAD', 'pyproject.toml'],
-          { cwd: tmpPluginDir }
-        );
-        // Read in the plugin's pyproject.toml, then delete it
-        const pyprojectTOML = toml.parse(fs.readFileSync(
-          upath.join(tmpPluginDir, 'pyproject.toml')
-        ).toString());
-        fs.rmSync(tmpPluginDir, { recursive: true, force: true });
+        let pyprojectTOML;
+        let installString;
+        if (url) { // install from git URL
+          if (revision) {
+            installString = `git+${url}@${revision}`;
+            logger.info(`adding plugin from ${installString}`);
+          } else {
+            installString = `git+${url}`;
+            logger.info(`adding plugin from ${installString} at default branch`);
+          }
 
+          // Create a temporary directory and check out the plugin's pyproject.toml,
+          // without downloading any extra files or git history
+          const tmpPluginDir = fs.mkdtempSync(upath.join(tmpdir(), 'natcap-invest-'));
+          await spawnWithLogging('git', ['clone', '--depth', 1, '--no-checkout', url, tmpPluginDir]);
+          let head = 'HEAD';
+          if (revision) {
+            head = 'FETCH_HEAD';
+            await spawnWithLogging(
+              'git',
+              ['fetch', 'origin', `${revision}`],
+              { cwd: tmpPluginDir }
+            );
+          }
+          await spawnWithLogging(
+            'git',
+            ['checkout', head, '--', 'pyproject.toml'],
+            { cwd: tmpPluginDir }
+          );
+          // Read in the plugin's pyproject.toml, then delete it
+          pyprojectTOML = toml.parse(fs.readFileSync(
+            upath.join(tmpPluginDir, 'pyproject.toml')
+          ).toString());
+          fs.rmSync(tmpPluginDir, { recursive: true, force: true });
+        } else { // install from local path
+          logger.info(`adding plugin from ${path}`);
+          // Read in the plugin's pyproject.toml, then delete it
+          pyprojectTOML = toml.parse(fs.readFileSync(
+            upath.join(path, 'pyproject.toml')
+          ).toString());
+        }
+
+        const micromamba = settingsStore.get('micromamba');
         // Access plugin metadata from the pyproject.toml
         const pluginID = pyprojectTOML.tool.natcap.invest.model_id;
         const pluginName = pyprojectTOML.tool.natcap.invest.model_name;
         const pluginPyName = pyprojectTOML.tool.natcap.invest.pyname;
-
         // Create a conda env containing the plugin and its dependencies
         const envName = `invest_plugin_${pluginID}`;
         await spawnWithLogging(
@@ -87,7 +110,7 @@ export function setupAddPlugin() {
         logger.info('created micromamba env for plugin');
         await spawnWithLogging(
           micromamba,
-          ['run', '--name', envName, 'pip', 'install', `git+${pluginURL}`]
+          ['run', '--name', envName, 'pip', 'install', installString]
         );
         logger.info('installed plugin into its env');
         // Write plugin metadata to the workbench's config.json
@@ -103,7 +126,7 @@ export function setupAddPlugin() {
             model_name: pluginName,
             pyname: pluginPyName,
             type: 'plugin',
-            source: pluginURL,
+            source: installString,
             env: envPath,
           }
         );
